CDP-312: S3 Upload & Transfer Middleware - Another monster commit. Moved transfer middleware to global scope. Modified transfer middelware to generically handle the download and upload of assets. Assets and retreived and updated via "models". Added abstract model that gets implemented by content models. Added content models for video and post. Modified AWS delete function to ignore requests that don't have amazonaws in the URL (unless the are relative paths and not URLs).

+++ src/services/amazon-aws.js
@@ -99,6 +99,7 @@ const upload = ( {
  */
 const remove = ( { url, bucket = process.env.AWS_S3_BUCKET } ) => {
   const args = URL.parse( url );
+  if ( args.hostname && args.hostname.indexOf( 'amazonaws' ) < 0 ) return;
   // Regex on path simply removes the preceeding '/' if any
   s3.deleteObject( { Key: args.path.replace( /^\/+(.*)$/g, '$1' ), Bucket: bucket }, ( err, data ) => {
     if ( err ) console.error( 'aws remove error', url, err );
